--- src/app/components/PromptCreateInputField/index.module.scss
@@ -1,29 +1,0 @@
-@use '../../style/lib.scss' as lib;
-
-.prompt-create-input-field {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-
-  padding: 30px 24px;
-}
-
-.button-group-container {
-  display: flex;
-  justify-content: flex-end;
-
-  width: 100%;
-}
-.button-group {
-  display: flex;
-  gap: 0.6rem;
-
-  @media screen and (max-width: 565px) {
-    width: 100%;
-  }
-}
-
-.propmt-button {
-  min-width: 87px;
-}
