+++ src/app/components/index.ts
@@ -4,6 +4,5 @@ export * from './ChatForm';
 export * from './ChatInput';
 export * from './ChatList';
 export * from './PromptCard';
-export * from './PromptCreateInputField';
-export * from './PromptCardGroup';
 export * from './PromptCreateForm';
+export * from './PromptCardGroup';
