+++ src/app/page.module.scss
@@ -8,3 +8,31 @@
   height: 100vh;
   background: #343541;
 }
+
+.form-placeholder {
+  color: rgba(86, 88, 105, 0.9);
+  font-size: 2rem;
+  font-weight: 600;
+  line-height: 2.5rem;
+  height: 100%;
+  padding-top: 4rem;
+}
+
+.prompt-create-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: auto;
+  gap: 1rem;
+  margin-bottom: 1rem;
+}
+
+.input-field-wrapper {
+  border: 1px solid rgb(86, 88, 105);
+  border-radius: 0.5rem;
+  width: 748px;
+
+  @media screen and (max-width: 760px) {
+    width: 100%;
+  }
+}
