+++ src/app/page.tsx
@@ -5,7 +5,14 @@ import { useChat } from 'ai/react';
 import { useMemo } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { PromptCreateForm, ChatList, ChatForm, Header } from './components';
+import {
+  ChatList,
+  ChatForm,
+  Header,
+  PromptCardGroup,
+  PromptCard,
+  PromptCreateForm,
+} from './components';
 import styles from './page.module.scss';
 
 const Home = () => {
@@ -15,8 +22,6 @@ const Home = () => {
     });
 
   const isChatInProgress = useMemo(() => !!messages.length, [messages.length]);
-
-  // TODO: message 폼 데이터 관리
   const methods = useForm({ defaultValues: { message: '' } });
 
   return (
@@ -26,7 +31,32 @@ const Home = () => {
         {isChatInProgress ? (
           <ChatList messages={messages} />
         ) : (
-          <PromptCreateForm setInput={setInput} />
+          <>
+            <h1 className={styles['form-placeholder']}>ChatGPT-MR</h1>
+            <div className={styles['prompt-create-container']}>
+              <div className={styles['input-field-wrapper']}>
+                <PromptCreateForm setInput={setInput} />
+              </div>
+            </div>
+            <PromptCardGroup>
+              <PromptCard
+                title="Come up with concepts"
+                description="for a retro-style arcade game"
+              />
+              <PromptCard
+                title="Recommend activities"
+                description="for a team-building day with remote employees"
+              />
+              <PromptCard
+                title="Show me a code snippet"
+                description="of a website's sticky header"
+              />
+              <PromptCard
+                title="Write a text message"
+                description="asking a friend to be my plus-one at a wedding"
+              />
+            </PromptCardGroup>
+          </>
         )}
         <ChatForm
           isChatInProgress={isChatInProgress}
